Annotation tool upgrade
With this commit a new signal and method were implemented in the
toolbar.py file, that keep track of the shape type that is selected
through the annotation toolbar. The method is needed to also keep
track of the drawing mode.

The is also a new method in the annotation_group, that basically
does the same.