Bug fix and tool tip
Fixed a bug where you could not draw any more after not committing to a shape and closing the label dialogue.
Tried to add the tool tip that one has to right click to end the annotation to the main widget tool tip. It kind of works, but the main_widget seems to need an update but that is also not really working.